refactor(jit): split jit.c into logical modules by responsibility
Split the 1946-line jit.c into focused modules:
- jit_internal.h: Shared header with extern declarations
- trap.c: Signal handling, stack bounds, alternate stack
- exec_mem.c: Executable memory allocation (mmap/VirtualAlloc)
- jit_context.c: JIT context allocation and management
- memory_ops.c: memory.grow/fill/copy, table.grow
- gc_type_cache.c: Type cache, ref.test/cast, subtype checking
- gc_ops.c: GC struct/array operations (new/get/set/len/fill/copy)
- jit.c: FFI exports only (reduced to 623 lines)
